separate base timer code into separate module
this way the pwm and the ppm module can share this code when the ppm is
ready
