+++ hal/stm32f373/Makefile
@@ -45,7 +45,7 @@ SRC-$(CONFIG_GPIO) += ./gpio.c
 SRC-$(CONFIG_SPIS) += ./spis.c
 SRC-$(CONFIG_SPIM) += ./spim.c
 SRC-$(CONFIG_ADC) += ./adc.c
-SRC-$(CONFIG_PWM) += ./pwm.c
+SRC-$(CONFIG_PWM) += ./pwm.c ./tmr.c
 
 ASRC = ./STM32F37x_DSP_StdPeriph_Lib_V1.0.0/Libraries/CMSIS/Device/ST/STM32F37x/Source/Templates/gcc_ride7/startup_stm32f37x.s
